Rename files according to convention; module organization
