+++ moneybot/markets/adapters/backtest.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 from typing import Iterator
 
-from moneybot.market_adapters import MarketAdapter
-from moneybot.MarketState import MarketState
-from moneybot.ProposedTrade import ProposedTrade
+from moneybot.markets.adapters import MarketAdapter
+from moneybot.markets.proposed_trade import ProposedTrade
+from moneybot.markets.state import MarketState
 
 
 class BacktestMarketAdapter(MarketAdapter):
