+++ moneybot/markets/state.py
