+++ moneybot/strategies/__init__.py
@@ -6,9 +6,9 @@ from typing import Generator
 from typing import List
 from typing import Set
 
-from moneybot.ProposedTrade import ProposedTrade
-from moneybot.MarketHistory import MarketHistory
-from moneybot.MarketState import MarketState
+from moneybot.markets.history import MarketHistory
+from moneybot.markets.proposed_trade import ProposedTrade
+from moneybot.markets.state import MarketState
 
 
 class Strategy(metaclass=ABCMeta):
