+++ moneybot/strategies/buffed_coin.py
